weather_service/requirements.txt
coverage==7.3.0
Django==4.2.4
django_environ==0.10.0
djangorestframework==3.14.0
psycopg2==2.9.7
requests==2.31.0